requirements.txt
flask==3.0.0
feedparser==6.0.10
requests==2.31.0
python-dateutil==2.8.2
anthropic==0.39.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
